chore: split out remaining plat-specific stuff
