+++ src/unix/CMakeLists.txt
@@ -28,10 +28,12 @@ if(WIN32)
     target_sources(plat PRIVATE
         ../qt/win_netsocket.c
         ../qt/win_dynld.c
+        sdl_plat_win.c
     )
     target_link_libraries(86Box ws2_32)
 else()
     target_sources(plat PRIVATE
+        sdl_plat_unix.c
         unix_netsocket.c
         unix_dynld.c
     )
