+++ src/unix/unix.c
@@ -1,66 +1,16 @@
-#ifdef __linux__
-#    define _FILE_OFFSET_BITS   64
-#    define _LARGEFILE64_SOURCE 1
-#endif
-#ifdef __HAIKU__
-#include <OS.h>
-#endif
-#include <SDL.h>
-#include <stdlib.h>
 #include <stdbool.h>
-#include <stdio.h>
-#include <string.h>
-#include <time.h>
-#include <sys/param.h>
-#include <sys/types.h>
-#include <sys/stat.h>
-#include <errno.h>
-#include <inttypes.h>
+#include <stdint.h>
 #include <wchar.h>
-#include <stdatomic.h>
 
-#ifdef _WIN32
-#    include <direct.h>
-#    include <windows.h>
-#else
-#    include <sys/time.h>
-#    include <sys/mman.h>
-#    include <sys/ioctl.h>
-#    include <fcntl.h>
-#    include <unistd.h>
-#    include <pwd.h>
-#    include <spawn.h>
-#endif
-
-/* Block device ioctl headers */
-#ifdef __linux__
-#    include <linux/fs.h>
-#endif
-#ifdef __APPLE__
-#    include <sys/disk.h>
-#endif
-#if defined(__FreeBSD__) || defined(__OpenBSD__) || defined(__NetBSD__)
-#    include <sys/disk.h>
-#    include <sys/disklabel.h>
-#endif
-
-#ifdef __APPLE__
-#    include "macOSXGlue.h"
-#endif
+#include <SDL.h>
 
 #include <86box/86box.h>
-#include <86box/mem.h>
-#include <86box/rom.h>
 #include <86box/keyboard.h>
 #include <86box/mouse.h>
-#include <86box/path.h>
 #include <86box/plat.h>
 #include <86box/thread.h>
-#include <86box/device.h>
-#include <86box/gameport.h>
 #include <86box/unix_sdl.h>
 #include <86box/unix_osd.h>
-#include "cpu.h"
 #include <86box/timer.h>
 #include <86box/nvr.h>
 #include <86box/version.h>
@@ -69,11 +19,6 @@
 #include <86box/gdbstub.h>
 
 #include "sdl_monitor.h"
-
-#ifndef _WIN32
-#    define __USE_GNU 1 /* shouldn't be done, yet it is */
-#    include <pthread.h>
-#endif
 
 extern SDL_Window         *sdl_win;
 
@@ -214,220 +159,6 @@ typedef struct sdl_blit_params {
 sdl_blit_params params  = { 0, 0, 0, 0 };
 int             blitreq = 0;
 
-#define TMPFILE_BUFSIZE 1024  // Assumed max buffer size
-void
-plat_tempfile(char *bufp, char *prefix, char *suffix)
-{
-#ifdef _WIN32
-    SYSTEMTIME t;
-    size_t     used = 0;
-
-    if (prefix != NULL)
-        used = snprintf(bufp, TMPFILE_BUFSIZE, "%s-", prefix);
-    else if (TMPFILE_BUFSIZE > 0)
-        bufp[0] = '\0';
-
-    GetLocalTime(&t);
-
-    if (used < TMPFILE_BUFSIZE) {
-        snprintf(bufp + used, TMPFILE_BUFSIZE - used,
-                 "%04u%02u%02u-%02u%02u%02u-%03u%s",
-                 (unsigned int) t.wYear, (unsigned int) t.wMonth,
-                 (unsigned int) t.wDay, (unsigned int) t.wHour,
-                 (unsigned int) t.wMinute, (unsigned int) t.wSecond,
-                 (unsigned int) t.wMilliseconds, suffix);
-    }
-#else
-    struct tm     *calendertime;
-    struct timeval t;
-    time_t         curtime;
-    size_t         used = 0;
-
-    if (prefix != NULL)
-        used = snprintf(bufp, TMPFILE_BUFSIZE, "%s-", prefix);
-    else if (TMPFILE_BUFSIZE > 0)
-        bufp[0] = '\0';
-
-    gettimeofday(&t, NULL);
-    curtime      = time(NULL);
-    calendertime = localtime(&curtime);
-
-    if (used < TMPFILE_BUFSIZE) {
-        snprintf(bufp + used, TMPFILE_BUFSIZE - used,
-                 "%d%02d%02d-%02d%02d%02d-%03" PRId32 "%s",
-                 calendertime->tm_year, calendertime->tm_mon, calendertime->tm_mday,
-                 calendertime->tm_hour, calendertime->tm_min, calendertime->tm_sec,
-                 (int32_t)(t.tv_usec / 1000), suffix);
-    }
-#endif
-}
-#undef TMPFILE_BUFSIZE
-
-
-#ifdef _WIN32
-size_t
-mbstoc16s(uint16_t dst[], const char src[], int len)
-{
-    if (src == NULL || len < 0)
-        return 0;
-
-    size_t ret = MultiByteToWideChar(CP_UTF8, 0, src, -1,
-                                     (LPWSTR) dst, dst == NULL ? 0 : len);
-
-    return ret ? ret : (size_t) -1;
-}
-
-size_t
-c16stombs(char dst[], const uint16_t src[], int len)
-{
-    if (src == NULL || len < 0)
-        return 0;
-
-    size_t ret = WideCharToMultiByte(CP_UTF8, 0, (LPCWCH) src, -1,
-                                     dst, dst == NULL ? 0 : len, NULL, NULL);
-
-    return ret ? ret : (size_t) -1;
-}
-
-void
-plat_get_system_directory(char *outbuf)
-{
-    if (outbuf == NULL)
-        return;
-
-    GetSystemWindowsDirectoryA(outbuf, MAX_PATH);
-    path_normalize(outbuf);
-}
-#endif
-
-int
-path_abs(char *path)
-{
-#ifdef _WIN32
-    return (path[1] == ':') || (path[0] == '\\') || (path[0] == '/') || (strstr(path, "ioctl://") == path);
-#else
-    return path[0] == '/';
-#endif
-}
-
-void
-path_normalize(char *path)
-{
-#ifdef _WIN32
-    if (strstr(path, "ioctl://") != path) {
-        while (*path != '\0') {
-            if (*path == '\\')
-                *path = '/';
-            path++;
-        }
-    }
-#else
-    (void) path;
-#endif
-}
-
-int
-plat_is_block_device(const char *path)
-{
-    struct stat stats;
-    if (stat(path, &stats) < 0)
-        return 0;
-#if defined(_WIN32)
-    return 0;
-#elif defined(__FreeBSD__) || defined(__OpenBSD__) || defined(__NetBSD__) || defined(__DragonFly__)
-    /* BSD systems use character devices for disk access, not block devices */
-    return S_ISCHR(stats.st_mode) ? 1 : 0;
-#else
-    return S_ISBLK(stats.st_mode) ? 1 : 0;
-#endif
-}
-
-int64_t
-plat_get_block_device_size(const char *path)
-{
-#ifdef __linux__
-    int fd = open(path, O_RDONLY);
-    if (fd < 0)
-        return -1;
-
-    uint64_t size = 0;
-    if (ioctl(fd, BLKGETSIZE64, &size) < 0) {
-        close(fd);
-        return -1;
-    }
-    close(fd);
-    return (int64_t) size;
-
-#elif defined(__APPLE__)
-    int fd = open(path, O_RDONLY);
-    if (fd < 0)
-        return -1;
-
-    uint64_t block_count = 0;
-    uint32_t block_size = 0;
-
-    if (ioctl(fd, DKIOCGETBLOCKCOUNT, &block_count) < 0) {
-        close(fd);
-        return -1;
-    }
-    if (ioctl(fd, DKIOCGETBLOCKSIZE, &block_size) < 0) {
-        close(fd);
-        return -1;
-    }
-
-    close(fd);
-    return (int64_t)(block_count * block_size);
-
-#elif defined(__FreeBSD__) || defined(__OpenBSD__) || defined(__NetBSD__)
-    int fd = open(path, O_RDONLY);
-    if (fd < 0)
-        return -1;
-
-    off_t size = 0;
-    if (ioctl(fd, DIOCGMEDIASIZE, &size) < 0) {
-        close(fd);
-        return -1;
-    }
-    close(fd);
-    return (int64_t) size;
-
-#else
-    /* Unsupported platform */
-    (void) path;
-    return -1;
-#endif
-}
-
-
-void *
-plat_mmap(size_t size, uint8_t executable)
-{
-#ifdef _WIN32
-    return VirtualAlloc(NULL, size, MEM_COMMIT | MEM_RESERVE,
-                        executable ? PAGE_EXECUTE_READWRITE : PAGE_READWRITE);
-#else
-#if defined __APPLE__ && defined MAP_JIT
-    void *ret = mmap(0, size, PROT_READ | PROT_WRITE | (executable ? PROT_EXEC : 0), MAP_ANON | MAP_PRIVATE | (executable ? MAP_JIT : 0), -1, 0);
-#elif defined(PROT_MPROTECT)
-    void *ret = mmap(0, size, PROT_MPROTECT(PROT_READ | PROT_WRITE | (executable ? PROT_EXEC : 0)), MAP_ANON | MAP_PRIVATE, -1, 0);
-#else
-    void *ret = mmap(0, size, PROT_READ | PROT_WRITE | (executable ? PROT_EXEC : 0), MAP_ANON | MAP_PRIVATE, -1, 0);
-#endif
-    return (ret == MAP_FAILED) ? NULL : ret;
-#endif
-}
-
-void
-plat_munmap(void *ptr, size_t size)
-{
-#ifdef _WIN32
-    (void) size;
-    VirtualFree(ptr, 0, MEM_RELEASE);
-#else
-    munmap(ptr, size);
-#endif
-}
-
 volatile int cpu_thread_run = 1;
 
 void
@@ -549,252 +280,6 @@ typedef struct mouseinputdata {
 SDL_mutex *mousemutex;
 int        real_sdl_w;
 int        real_sdl_h;
-
-#define TMP_PATH_BUFSIZE 1024
-
-static int
-build_subdir_path(char *outbuf, size_t len, const char *base_path, const char *subdir)
-{
-    int written = snprintf(outbuf, len, "%s%s", base_path, subdir);
-
-    if ((written < 0) || ((size_t) written >= len)) {
-        if (len > 0)
-            outbuf[0] = '\0';
-        return 0;
-    }
-
-    return 1;
-}
-
-void
-plat_init_rom_paths(void)
-{
-#ifdef _WIN32
-    char global_data_path[TMP_PATH_BUFSIZE] = {0};
-    char global_rom_path[TMP_PATH_BUFSIZE]  = {0};
-
-    plat_get_global_data_dir(global_data_path, sizeof(global_data_path));
-    if (build_subdir_path(global_rom_path, sizeof(global_rom_path), global_data_path, "roms")) {
-        if (!plat_dir_check(global_rom_path))
-            plat_dir_create(global_rom_path);
-        rom_add_path(global_rom_path);
-    }
-
-    rom_add_path("roms/");
-#else
-#ifndef __APPLE__
-    const char *xdg_data_home = getenv("XDG_DATA_HOME");
-    if (xdg_data_home) {
-        char xdg_rom_path[TMP_PATH_BUFSIZE] = {0};
-        size_t used = snprintf(xdg_rom_path, sizeof(xdg_rom_path), "%s/", xdg_data_home);
-        if (used < sizeof(xdg_rom_path))
-            used += snprintf(xdg_rom_path + used, sizeof(xdg_rom_path) - used, EMU_NAME "/");
-        if (used < sizeof(xdg_rom_path) && !plat_dir_check(xdg_rom_path))
-            plat_dir_create(xdg_rom_path);
-        if (used < sizeof(xdg_rom_path))
-            used += snprintf(xdg_rom_path + used, sizeof(xdg_rom_path) - used, "roms/");
-        if (used < sizeof(xdg_rom_path) && !plat_dir_check(xdg_rom_path))
-            plat_dir_create(xdg_rom_path);
-        if (used < sizeof(xdg_rom_path))
-            rom_add_path(xdg_rom_path);
-    } else {
-        const char *home = getenv("HOME");
-        if (!home) {
-            struct passwd *pw = getpwuid(getuid());
-            if (pw)
-                home = pw->pw_dir;
-        }
-
-        if (home) {
-            char home_rom_path[TMP_PATH_BUFSIZE] = {0};
-            size_t used = snprintf(home_rom_path, sizeof(home_rom_path),
-                                   "%s/.local/share/" EMU_NAME "/", home);
-            if (used < sizeof(home_rom_path) && !plat_dir_check(home_rom_path))
-                plat_dir_create(home_rom_path);
-            if (used < sizeof(home_rom_path))
-                used += snprintf(home_rom_path + used,
-                                 sizeof(home_rom_path) - used, "roms/");
-            if (used < sizeof(home_rom_path) && !plat_dir_check(home_rom_path))
-                plat_dir_create(home_rom_path);
-            if (used < sizeof(home_rom_path))
-                rom_add_path(home_rom_path);
-        }
-    }
-
-    const char *xdg_data_dirs = getenv("XDG_DATA_DIRS");
-    if (xdg_data_dirs) {
-        char *xdg_rom_paths = strdup(xdg_data_dirs);
-        if (xdg_rom_paths) {
-            // Trim trailing colons
-            size_t len = strlen(xdg_rom_paths);
-            while (len > 0 && xdg_rom_paths[len - 1] == ':')
-                xdg_rom_paths[--len] = '\0';
-
-            char *saveptr = NULL;
-            char *cur_xdg = strtok_r(xdg_rom_paths, ":", &saveptr);
-            while (cur_xdg) {
-                char real_xdg_rom_path[TMP_PATH_BUFSIZE] = {0};
-                size_t used = snprintf(real_xdg_rom_path,
-                                       sizeof(real_xdg_rom_path),
-                                       "%s/" EMU_NAME "/roms/", cur_xdg);
-                if (used < sizeof(real_xdg_rom_path))
-                    rom_add_path(real_xdg_rom_path);
-                cur_xdg = strtok_r(NULL, ":", &saveptr);
-            }
-
-            free(xdg_rom_paths);
-        }
-    } else {
-        rom_add_path("/usr/local/share/" EMU_NAME "/roms/");
-        rom_add_path("/usr/share/" EMU_NAME "/roms/");
-    }
-#else
-    char default_rom_path[TMP_PATH_BUFSIZE] = {0};
-    getDefaultROMPath(default_rom_path);
-    rom_add_path(default_rom_path);
-#endif
-#endif
-}
-
-void
-plat_init_asset_paths(void)
-{
-#ifdef _WIN32
-    char global_data_path[TMP_PATH_BUFSIZE] = {0};
-    char global_asset_path[TMP_PATH_BUFSIZE] = {0};
-
-    plat_get_global_data_dir(global_data_path, sizeof(global_data_path));
-    if (build_subdir_path(global_asset_path, sizeof(global_asset_path), global_data_path, "assets")) {
-        if (!plat_dir_check(global_asset_path))
-            plat_dir_create(global_asset_path);
-        asset_add_path(global_asset_path);
-    }
-
-    asset_add_path("assets/");
-#else
-#ifndef __APPLE__
-    const char *xdg_data_home = getenv("XDG_DATA_HOME");
-    if (xdg_data_home) {
-        char xdg_asset_path[TMP_PATH_BUFSIZE] = {0};
-        size_t used = snprintf(xdg_asset_path, sizeof(xdg_asset_path), "%s/", xdg_data_home);
-        if (used < sizeof(xdg_asset_path))
-            used += snprintf(xdg_asset_path + used, sizeof(xdg_asset_path) - used, EMU_NAME "/");
-        if (used < sizeof(xdg_asset_path) && !plat_dir_check(xdg_asset_path))
-            plat_dir_create(xdg_asset_path);
-        if (used < sizeof(xdg_asset_path))
-            used += snprintf(xdg_asset_path + used, sizeof(xdg_asset_path) - used, "assets/");
-        if (used < sizeof(xdg_asset_path) && !plat_dir_check(xdg_asset_path))
-            plat_dir_create(xdg_asset_path);
-        if (used < sizeof(xdg_asset_path))
-            asset_add_path(xdg_asset_path);
-    } else {
-        const char *home = getenv("HOME");
-        if (!home) {
-            struct passwd *pw = getpwuid(getuid());
-            if (pw)
-                home = pw->pw_dir;
-        }
-
-        if (home) {
-            char home_asset_path[TMP_PATH_BUFSIZE] = {0};
-            size_t used = snprintf(home_asset_path, sizeof(home_asset_path),
-                                   "%s/.local/share/" EMU_NAME "/", home);
-            if (used < sizeof(home_asset_path) && !plat_dir_check(home_asset_path))
-                plat_dir_create(home_asset_path);
-            if (used < sizeof(home_asset_path))
-                used += snprintf(home_asset_path + used,
-                                 sizeof(home_asset_path) - used, "assets/");
-            if (used < sizeof(home_asset_path) && !plat_dir_check(home_asset_path))
-                plat_dir_create(home_asset_path);
-            if (used < sizeof(home_asset_path))
-                asset_add_path(home_asset_path);
-        }
-    }
-
-    const char *xdg_data_dirs = getenv("XDG_DATA_DIRS");
-    if (xdg_data_dirs) {
-        char *xdg_asset_paths = strdup(xdg_data_dirs);
-        if (xdg_asset_paths) {
-            // Trim trailing colons
-            size_t len = strlen(xdg_asset_paths);
-            while (len > 0 && xdg_asset_paths[len - 1] == ':')
-                xdg_asset_paths[--len] = '\0';
-
-            char *saveptr = NULL;
-            char *cur_xdg = strtok_r(xdg_asset_paths, ":", &saveptr);
-            while (cur_xdg) {
-                char real_xdg_asset_path[TMP_PATH_BUFSIZE] = {0};
-                size_t used = snprintf(real_xdg_asset_path,
-                                       sizeof(real_xdg_asset_path),
-                                       "%s/" EMU_NAME "/assets/", cur_xdg);
-                if (used < sizeof(real_xdg_asset_path))
-                    asset_add_path(real_xdg_asset_path);
-                cur_xdg = strtok_r(NULL, ":", &saveptr);
-            }
-
-            free(xdg_asset_paths);
-        }
-    } else {
-        asset_add_path("/usr/local/share/" EMU_NAME "/assets/");
-        asset_add_path("/usr/share/" EMU_NAME "/assets/");
-    }
-#else
-    char default_asset_path[TMP_PATH_BUFSIZE] = {0};
-    getDefaultROMPath(default_asset_path);
-    asset_add_path(default_asset_path);
-#endif
-#endif
-}
-#undef TMP_PATH_BUFSIZE
-
-void
-plat_get_global_config_dir(char *outbuf, const size_t len)
-{
-    return plat_get_global_data_dir(outbuf, len);
-}
-
-void
-plat_get_global_data_dir(char *outbuf, const size_t len)
-{
-    if (len == 0)
-        return;
-
-    if (portable_mode) {
-        strncpy(outbuf, exe_path, len);
-    } else {
-#ifdef _WIN32
-        const char *appdata = getenv("LOCALAPPDATA");
-
-        if (!appdata || !build_subdir_path(outbuf, len, appdata, "/" EMU_NAME))
-            strncpy(outbuf, exe_path, len);
-        else if (!plat_dir_check(outbuf))
-            plat_dir_create(outbuf);
-#else
-        char *prefPath = SDL_GetPrefPath(NULL, EMU_NAME);
-
-        if (prefPath != NULL) {
-            strncpy(outbuf, prefPath, len);
-            SDL_free(prefPath);
-        } else {
-            strncpy(outbuf, exe_path, len);
-        }
-#endif
-    }
-
-    outbuf[len - 1] = '\0';
-    path_slash(outbuf);
-}
-
-void
-plat_get_temp_dir(char *outbuf, uint8_t len)
-{
-    const char *tmpdir = getenv("TMPDIR");
-    if (tmpdir == NULL) {
-        tmpdir = "/tmp";
-    }
-    strncpy(outbuf, tmpdir, len);
-    path_slash(outbuf);
-}
 
 uint32_t
 timer_onesec(uint32_t interval, UNUSED(void *param))
@@ -1118,122 +603,6 @@ WinMain(HINSTANCE instance, HINSTANCE previous, LPSTR command_line, int show)
     return main(__argc, __argv);
 }
 #endif
-
-void
-plat_set_thread_name(void *thread, const char *name)
-{
-#ifdef _WIN32
-    (void) thread;
-    (void) name;
-#else
-#ifdef __APPLE__
-    if (thread) /* macOS pthread can only set self's name */
-        return;
-    char truncated[64];
-#elif defined(__NetBSD__)
-    char truncated[64];
-#elif defined(__HAIKU__)
-    char truncated[32];
-#else
-    char truncated[16];
-#endif
-    strncpy(truncated, name, sizeof(truncated) - 1);
-#ifdef __APPLE__
-    pthread_setname_np(truncated);
-#elif defined(__NetBSD__)
-    pthread_setname_np(thread ? *((pthread_t *) thread) : pthread_self(), truncated, "%s");
-#elif defined(__HAIKU__)
-    rename_thread(find_thread(NULL), truncated);
-#else
-    pthread_setname_np(thread ? *((pthread_t *) thread) : pthread_self(), truncated);
-#endif
-#endif
-}
-
-int
-plat_run_command(const char *cmd, const char **env, const char *title)
-{
-#ifdef _WIN32
-    (void) env;
-    (void) title;
-    return system(cmd) == 0;
-#else
-    /* Append environment variables to the existing environment. */
-    extern char **environ;
-    const  char **new_env = NULL;
-    if (env && env[0] && env[0][0]) {
-        int c = 0;
-        while (environ[c])
-            c++;
-        int d = 0;
-        while (env[d] && env[d][0])
-            d++;
-        new_env = (const char **) malloc((c + d + 1) * sizeof(char *));
-        for (c = 0; environ[c]; c++)
-            new_env[c] = environ[c];
-        for (d = 0; env[d] && env[d][0]; d++)
-            new_env[c++] = env[d];
-        new_env[c] = NULL;
-    }
-
-    /* Execute command under a terminal emulator if requested. */
-    int         ret;
-    pid_t       pid;
-    const char *args[] = {NULL, NULL, NULL, NULL, "/bin/sh", "-c", cmd, NULL};
-    if (title) {
-        /* Set arguments for xdg-terminal-exec. */
-        int   len       = (strlen(title) * 3) + 9;
-        char *title_buf = (char *) malloc(len);
-        int   c         = snprintf(title_buf, len, "--title=");
-        char *title_val = &title_buf[c];
-        int   bracket   = 0;
-        for (int i = 0; title[i]; i++) {
-            if (title[i] == '\n') {
-                if (bracket)
-                    title_buf[c++] = ']';
-                else
-                    bracket = 1;
-                title_buf[c++] = ' ';
-                title_buf[c++] = '[';
-            } else {
-                title_buf[c++] = title[i];
-            }
-        }
-        if (bracket)
-            title_buf[c++] = ']';
-        title_buf[c] = '\0';
-        args[1]      = title_buf;
-        len          = strlen(usr_path) + 7;
-        args[2]      = malloc(len);
-        snprintf((char *) args[2], len, "--dir=%s", usr_path);
-        args[3] = "--";
-
-        /* Try executing terminals. */
-        ret = 0;
-        static const char *terminals[] = {"xdg-terminal-exec", "x-terminal-emulator", "xterm", "urxvt", "rxvt"};
-        for (int i = 0; (i < (sizeof(terminals) / sizeof(terminals[0]))) && !ret; i++) {
-            args[0] = terminals[i];
-            ret     = !posix_spawnp(&pid, args[0], NULL, NULL, (char * const *) args, new_env ? (char * const *) new_env : (char * const *) environ);
-            if (len) {
-                /* Set arguments for other terminals. */
-                len = 0;
-                free((void *) args[2]);
-                args[1] = "-T";
-                args[2] = title_val;
-                args[3] = "-e";
-            }
-        }
-        free(title_buf);
-    } else {
-        /* Execute command directly. */
-        ret = !posix_spawn(&pid, args[4], NULL, NULL, (char * const *) &args[4], new_env ? (char * const *) new_env : (char * const *) environ);
-    }
-
-    if (new_env)
-        free(new_env);
-    return ret;
-#endif
-}
 
 void
 startblit(void)
